changed hpp to h

+++ backups/bak_src/ray.cpp
@@ -1,4 +1,4 @@
-#include "ray.hpp"
+#include "ray.h"
 
 const point3& ray::origin() const  { return origin_; }
 const vec3& ray::direction() const { return direction_; }
@@ -7,7 +7,7 @@ ray::ray()
 {}
 ray::ray(const point3& origin, const vec3& direction) : origin_(origin), direction_(direction)
 {}
-point3 ray::at(double t) const {
+point3 ray::at(float t) const {
     return origin_ + t*direction_;
 }
